Add WebCommon shared component kit + NuGet workflow
A small Blazor RCL (published as RaidMax.IW4MAdmin.WebCommon) so plugin pages
reuse the host's native UI instead of rolling their own navigation/layout.

- SideContextMenu moved out of WebfrontCore into WebCommon. Its AppState
  dependency is replaced by an optional Localizer delegate (host passes
  AppState.Loc); the mobile drawer is reworked to be CSS-only (hidden checkbox
  + peer-checked: + <label> toggles) so it works on statically-rendered plugin
  pages too, with no Blazor interactivity required.
- PluginPageShell: the standard content + sticky right-pane page layout, so a
  plugin page matches a first-party stats/leaderboard page.
- Host Tailwind input gains @source "../../../../WebCommon" so the shared
  components' classes are compiled into the host stylesheet (loaded on every
  page); plugin pages render with host classes and never depend on their own
  Tailwind build for layout.
- All first-party SideContextMenu consumers migrated; the three collapse-group
  sites pass Localizer=AppState.Loc.
- CI: the SharedLibraryCore NuGet workflow now also builds/packs/publishes
  WebCommon, version-locked to SharedLibraryCore.

The bundle targets mark WebCommon.dll host-provided so plugins never bundle it.
