+++ WebCommon/Components/UI/Navigation/SideContextMenu.razor
@@ -1,30 +1,43 @@
-@* Mobile FAB - only visible below xl breakpoint *@
+@namespace WebCommon.Components.UI.Navigation
+
+@* Right-aligned context navigation, shared by the host and by plugins.
+   Desktop: a sticky right-hand pane (xl and up).
+   Mobile: a CSS-only drawer driven by a hidden checkbox + <label> toggles, so it works on
+   statically-rendered (SSR) plugin pages too — no Blazor interactivity required.
+   Localization is injected via the optional Localizer delegate (host passes AppState.Loc;
+   plugins can omit it and group headers fall back to their raw key). *@
+
 @if (Model?.Items?.Any() == true)
 {
-    <button type="button" 
-            @onclick="ToggleMobileMenu"
-            @onclick:stopPropagation="true"
-            class="fixed bottom-6 right-6 z-[60] xl:hidden w-14 h-14 rounded-full shadow-2xl 
-                   @(_isOpen ? "bg-secondary" : "bg-primary") text-white
-                   active:scale-95 
-                   focus:outline-none focus:ring-4 focus:ring-primary/30
-                   transition-all duration-300 ease-out
-                   flex items-center justify-center">
-        <i class="ph @(_isOpen ? "ph-x" : "ph-list") text-2xl transition-transform duration-200 @(_isOpen ? "rotate-90" : "rotate-0")"></i>
-    </button>
+    @* ---- Mobile (below xl): CSS-only drawer ------------------------------------------------ *@
+    <div class="xl:hidden">
+        <input type="checkbox" id="scm-mobile-toggle" class="peer hidden" />
 
-    @* Mobile Drawer *@
-    <div class="fixed inset-0 z-[55] xl:hidden @(_isOpen ? "pointer-events-auto" : "pointer-events-none")">
-        @* Overlay with fade animation *@
-        <div class="fixed inset-0 bg-black/60 backdrop-blur-sm transition-opacity duration-300 @(_isOpen ? "opacity-100" : "opacity-0")" 
-             @onclick="ToggleMobileMenu"></div>
-        
-        @* Drawer Panel with slide animation *@
-        <div class="fixed right-0 top-0 h-full w-80 max-w-[85vw] bg-surface/95 backdrop-blur-xl border-l border-line/50 shadow-2xl 
-                    overflow-y-auto transform transition-transform duration-300 ease-out
-                    @(_isOpen ? "translate-x-0" : "translate-x-full")">
-            
-            @* Drawer Header *@
+        @* open FAB (hidden once the drawer is open) *@
+        <label for="scm-mobile-toggle"
+               class="fixed bottom-6 right-6 z-[60] w-14 h-14 rounded-full shadow-2xl bg-primary text-white
+                      active:scale-95 transition-all duration-300 ease-out flex items-center justify-center
+                      cursor-pointer peer-checked:hidden">
+            <i class="ph ph-list text-2xl"></i>
+        </label>
+
+        @* close FAB (shown only while open) *@
+        <label for="scm-mobile-toggle"
+               class="fixed bottom-6 right-6 z-[60] w-14 h-14 rounded-full shadow-2xl bg-secondary text-white
+                      active:scale-95 transition-all duration-300 ease-out hidden items-center justify-center
+                      cursor-pointer peer-checked:flex">
+            <i class="ph ph-x text-2xl"></i>
+        </label>
+
+        @* overlay *@
+        <label for="scm-mobile-toggle"
+               class="fixed inset-0 z-[55] bg-black/60 backdrop-blur-sm opacity-0 pointer-events-none
+                      transition-opacity duration-300 peer-checked:opacity-100 peer-checked:pointer-events-auto"></label>
+
+        @* drawer panel *@
+        <div class="fixed right-0 top-0 z-[56] h-full w-80 max-w-[85vw] bg-surface/95 backdrop-blur-xl
+                    border-l border-line/50 shadow-2xl overflow-y-auto transform translate-x-full
+                    transition-transform duration-300 ease-out peer-checked:translate-x-0">
             <div class="sticky top-0 z-10 bg-surface/90 backdrop-blur-lg border-b border-line/50 px-5 py-4">
                 <div class="flex items-center justify-between">
                     <div class="flex items-center gap-3">
@@ -33,23 +46,21 @@
                         </div>
                         <h3 class="text-sm font-bold text-foreground uppercase tracking-wider">@Model.MenuTitle</h3>
                     </div>
-                    <button type="button" @onclick="ToggleMobileMenu" 
-                            class="w-8 h-8 rounded-full bg-surface-alt hover:bg-line flex items-center justify-center text-muted hover:text-foreground transition-all duration-200">
+                    <label for="scm-mobile-toggle"
+                           class="w-8 h-8 rounded-full bg-surface-alt hover:bg-line flex items-center justify-center text-muted hover:text-foreground transition-all duration-200 cursor-pointer">
                         <i class="ph ph-x text-lg"></i>
-                    </button>
+                    </label>
                 </div>
             </div>
-            
-            @* Drawer Content *@
+
             <div class="p-4">
                 @RenderMobileMenuContent()
             </div>
         </div>
     </div>
 
-@* Desktop Sidebar - only visible at xl breakpoint and above *@
-    <div
-        class="bg-surface border-l border-y border-line w-72 flex-shrink-0 hidden xl:block p-4 rounded-l-xl max-h-[calc(100vh-3rem)] overflow-y-auto ml-auto">
+    @* ---- Desktop (xl and up): sticky right pane ------------------------------------------- *@
+    <div class="bg-surface border-l border-y border-line w-72 flex-shrink-0 hidden xl:block p-4 rounded-l-xl max-h-[calc(100vh-3rem)] overflow-y-auto ml-auto">
         @RenderMenuContent()
     </div>
 }
@@ -62,7 +73,6 @@
                 var groupedItems = Model.Items.Where(item => item.IsCollapse)
                     .GroupBy(item => item.Meta)
                     .ToList();
-                var index = 0;
             }
 
             @foreach (var item in Model.Items.Where(item => !item.IsCollapse))
@@ -97,7 +107,7 @@
                 <details class="group mt-4" open>
                     <summary
                         class="list-none flex items-center justify-between text-xs font-bold text-muted uppercase tracking-wider mb-2 cursor-pointer hover:text-foreground transition-colors px-2">
-                        <span>@(AppState.Loc($"GAME_{group.Key}"))</span>
+                        <span>@Localize($"GAME_{group.Key}")</span>
                         <i class="ph ph-caret-down ml-2 transition-transform duration-200 group-open:rotate-180"></i>
                     </summary>
 
@@ -125,7 +135,6 @@
                         }
                     </div>
                 </details>
-                index++;
             }
         </div>
     };
@@ -137,7 +146,6 @@
                 var groupedItems = Model.Items.Where(item => item.IsCollapse)
                     .GroupBy(item => item.Meta)
                     .ToList();
-                var index = 0;
             }
 
             @foreach (var item in Model.Items.Where(item => !item.IsCollapse))
@@ -169,7 +177,7 @@
                 <details class="group mt-4" open>
                     <summary
                         class="list-none flex items-center justify-between text-xs font-bold text-muted uppercase tracking-wider mb-2 cursor-pointer hover:text-foreground transition-colors px-2">
-                        <span>@(AppState.Loc($"GAME_{group.Key}"))</span>
+                        <span>@Localize($"GAME_{group.Key}")</span>
                         <i class="ph ph-caret-down ml-2 transition-transform duration-200 group-open:rotate-180"></i>
                     </summary>
 
@@ -194,7 +202,6 @@
                         }
                     </div>
                 </details>
-                index++;
             }
         </div>
     };
