+++ WebCommon/_Imports.razor
@@ -1,0 +1,12 @@
+@using System
+@using System.Collections.Generic
+@using System.Linq
+@using System.Threading.Tasks
+
+@using Microsoft.AspNetCore.Components
+@using Microsoft.AspNetCore.Components.Web
+
+@using SharedLibraryCore.Dtos
+
+@using WebCommon.Components.UI.Navigation
+@using WebCommon.Components.UI.Layout
