+++ WebfrontCore/Components/Features/Clients/Statistics/AdvancedStats.razor
@@ -362,7 +362,7 @@
         @* ================================================================ *@
         <div class="xl:w-80 xl:shrink-0 xl:py-6 xl:pr-0 pl-0">
             <div class="sticky top-6">
-                <SideContextMenu Model="MenuItems"/>
+                <SideContextMenu Model="MenuItems" Localizer="AppState.Loc"/>
             </div>
         </div>
     }
