+++ WebfrontCore/Components/Features/Clients/Statistics/StatsOverview.razor
@@ -211,7 +211,7 @@
         @if (MenuItems != null)
         {
             <div class="sticky top-6">
-                <SideContextMenu Model="MenuItems"/>
+                <SideContextMenu Model="MenuItems" Localizer="AppState.Loc"/>
             </div>
         }
     </div>
