+++ WebfrontCore/Components/Features/Servers/Pages/ServerScoreboard.razor
@@ -24,7 +24,7 @@
         @if (ContextItems != null)
         {
            <div class="sticky top-6">
-               <SideContextMenu Model="ContextItems" />
+               <SideContextMenu Model="ContextItems" Localizer="AppState.Loc" />
            </div>
         }
     </div>
