+++ WebfrontCore/Components/_Imports.razor
@@ -56,6 +56,8 @@
 @using WebfrontCore.Components.UI.Layout
 @using WebfrontCore.Components.UI.Navigation
 @using WebfrontCore.Components.UI.Controls
+@using WebCommon.Components.UI.Navigation
+@using WebCommon.Components.UI.Layout
 @using WebfrontCore.Components.UI
 @using WebfrontCore.Components.UI.Display
 @using WebfrontCore.Core.Services
